requirements.txt
djangorestframework
django-cors-headers
networkx
graphviz
pandas
matplotlib
openpyxl
natsort
pydot
